+++ packages/superdough/helpers.mjs
@@ -1,4 +1,4 @@
-import { getAudioContext } from './superdough.mjs';
+import { getAudioContext } from './webaudio.mjs';
 
 export function gainNode(value) {
   const node = getAudioContext().createGain();
